requirements.txt
flask==2.0.1
google-cloud-storage==2.1.0
gunicorn==20.1.0